// src/app/generated/model/itemCreateRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ItemCreateRequest { 
    /**
     * Unique stock keeping unit code for the item
     */
    sku: string;
    /**
     * Display name of the item
     */
    name: string;
    /**
     * Category the item belongs to
     */
    category?: string;
    /**
     * Unit of measure used for stock counts
     */
    unitOfMeasure?: string;
    /**
     * Price at which the item is sold, must be non-negative
     */
    sellingPrice?: number;
    /**
     * Optional reference-only default cost price, must be non-negative
     */
    defaultCostPrice?: number;
}